+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod download;
 pub mod encoder;
+pub mod history;
 pub mod log;
 pub mod session;
 pub mod stream;